requirements.txt
google-api-python-client>=2.157.0,<3.0.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-storage==2.18.2
requests>=2.32.4,<3.0.0
google-generativeai==0.8.2
python-dotenv==1.0.1
json5==0.12.0
PyYAML>=6.0.1,<7.0.0
flask>=3.0
